Revert "Lab 3 def"
