+++ lab_01/data.txt
@@ -1,0 +1,9 @@
+8 3
+0.5 -0.301 0.867
+1 0 0.434
+1.5 0.176 0.289
+2 0.301 0.217
+2.5 0.398 0.174
+3 0.477 0.145
+3.5 0.544 0.124
+4 0.602 0.109
