+++ lab_01/data1.txt
@@ -1,0 +1,9 @@
+8 3
+0.00 1.000000 -1.000000
+0.15 0.838771 -1.14944
+0.30 0.655336 -1.29552
+0.45 0.450447 -1.43497
+0.60 0.225336 -1.56464
+0.75 -0.018310 -1.68164
+0.90 -0.278390 -1.78333
+1.05 -0.552430 -1.86742
