+++ lab_01/data2.txt
@@ -1,0 +1,10 @@
+9 3
+-6.28 1 0
+-4.71 0 1
+-3.14 1 0
+-1.57 0 1
+0 1 0
+1.57 0 1
+3.14 1 0
+4.71 0 1
+6.28 1 0
