+++ lab_02/CMakeLists.txt
@@ -1,0 +1,23 @@
+project(ca)
+
+cmake_minimum_required(VERSION 3.21)
+
+set(CMAKE_CXX_STANDARD 20)
+
+include_directories(
+    inc
+)
+
+add_executable(lab_02
+        src/main.cpp
+        inc/Matrix.h
+        src/Menu.cpp
+        inc/Menu.h
+        src/interpolate.cpp
+        inc/interpolate.h
+        src/MyErrors.cpp
+        inc/MyErrors.h
+        inc/CppConsoleTable.hpp
+        src/Matrix.cpp
+        src/Matrix3x.cpp
+        inc/Matrix3x.h)
