+++ lab_02/data.txt
@@ -1,0 +1,37 @@
+5
+6 6
+
+0	0	1	2	3	4
+0	0	1	4	9	16
+1	1	2	5	10	17
+2	4	5	8	13	20
+3	9	10	13	18	25
+4	16	17	20	25	32
+
+1	0	1	2	3	4
+0	1	2	5	10	17
+1	2	3	6	11	18
+2	5	6	9	14	21
+3	10	11	14	19	26
+4	17	18	21	26	33
+
+2	0	1	2	3	4
+0	4	5	8	13	20
+1	5	6	9	14	21
+2	8	9	12	17	24
+3	13	14	17	22	29
+4	20	21	24	29	36
+
+3	0	1	2	3	4
+0	9	10	13	18	25
+1	10	11	14	19	26
+2	13	14	17	22	29
+3	18	19	22	27	34
+4	25	26	29	34	41
+
+4	0	1	2	3	4
+0	16	17	20	25	32
+1	17	18	21	26	33
+2	20	21	24	29	36
+3	25	26	29	34	41
+4	32	33	36	41	48
